pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-cloud-storage>=3.3.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]

[tool.uv]
link-mode = "copy"
